Created the Difficulty Selection View
This view is for selecting the difficulty of the AI or to play with a
friend/actor (be it a human, dog, alien, idc)

It contains 4 buttons for the 3 difficulties, and 1 for selecting an
actor. The buttons are arranged in a square pattern. Each button calls
TheForce.selectDifficulty() when pressed with a variable passed to set
the difficulty of the AI or disable the AI then hide the difficulty
selection view.