Created the "who goes first" view
Created the “who goes first” view.

This view prompts the user to select which player should go first. It
displays a blue stone with the name of the user or “Guest”, and a red
stone with name “Player 2”. The stones then call
TheForce.whoGoesFirst() passing the value of the player selected, which
will set the player and hide the “who goes first” view.