Menus, Dialog, Undo, Formatting. See details.
Fixed some bugs regarding menusOpen returning the wrong value.

Added slot for onSetPlayerFinished, this will show a dialog prompting
the user if they want to continue as the username they typed.

Changed the turn making to actually call a function to show the win
dialog.

Removed the undo button, since we removed that from the specs.

Added a dialog specifically for displaying winning message, and a
dialog that’s generic and can be used for anything.