Created dialogs, disable/enable functions, etc
Login and Register are now fully implemented with the login menu.

Created multiple dialogs for login and register results.

Created enable and disable function to disable buttons in the
background when dialogs are visible

Created noInternet function to disable login and register when the user
is not connected to the internet. (Not yet implemented)